Refactor code to add HTTP handlers for "/hello" and "/form" routes, and serve static files
This commit refactors the code in main.go to add two new HTTP handlers for the "/hello" and "/form" routes. It also includes the necessary logic to serve static files using the http.FileServer. Additionally, two new HTML files, form.html and index.html, are added to the static directory.